Move profile api methods into their own module
